fixed add gig form validation error
The callback for the check_dropdown() method was entered in camel case when method is actually named with an undercase.

Also, in order to avoid database errors I added a zero on line 128 in Gig_model.php. This is so the 'id' column isn't taking in a null value. I'm guessing this would actually need the employer's own unique id number when which would be passed to the variable $userID when they login.